Change TopicData offset to Long from Double

+++ oura-library/src/main/kotlin/org/radarbase/oura/converter/OuraDailyActivityClassConverter.kt
@@ -35,7 +35,7 @@ class OuraDailyActivityClassConverter(
                 TopicData(
                     key = user.observationKey,
                     topic = topic,
-                    offset = time,
+                    offset = time.toLong(),
                     value =
                     toActivityClass(
                         time,
