+++ oura-library/src/main/kotlin/org/radarbase/oura/converter/OuraHeartRateConverter.kt
@@ -24,7 +24,7 @@ class OuraHeartRateConverter(
                 TopicData(
                     key = user.observationKey,
                     topic = topic,
-                    offset = startInstant.toEpochMilli().toDouble(),
+                    offset = startInstant.toEpochMilli(),
                     value = it.toHeartRate(startInstant),
                 )
             }
