+++ oura-library/src/main/kotlin/org/radarbase/oura/converter/OuraSessionConverter.kt
@@ -25,7 +25,7 @@ class OuraSessionConverter(
                 TopicData(
                     key = user.observationKey,
                     topic = topic,
-                    offset = startInstant.toEpochMilli().toDouble(),
+                    offset = startInstant.toEpochMilli(),
                     value = it.toSession(startInstant),
                 )
             }
