+++ oura-library/src/main/kotlin/org/radarbase/oura/converter/OuraSessionHeartRateConverter.kt
@@ -44,7 +44,7 @@ class OuraSessionHeartRateConverter(
                     TopicData(
                         key = user.observationKey,
                         topic = topic,
-                        offset = time,
+                        offset = time.toLong(),
                         value = toHeartRate(
                             time,
                             timeReceivedEpoch,
