+++ oura-library/src/main/kotlin/org/radarbase/oura/converter/OuraTagConverter.kt
@@ -30,7 +30,7 @@ class OuraTagConverter(
                             TopicData(
                                 key = user.observationKey,
                                 topic = topic,
-                                offset = startInstant.toEpochMilli().toDouble(),
+                                offset = startInstant.toEpochMilli(),
                                 value = data.toTag(startInstant, it.textValue()),
                             )
                         }
